Vec3 -> Vec3d

+++ Minebot/src/net/famzangl/minecraft/minebot/ai/path/world/BlockBounds.java
@@ -3,6 +3,7 @@ package net.famzangl.minecraft.minebot.ai.path.world;
 import net.famzangl.minecraft.minebot.ai.utils.RandUtils;
 import net.minecraft.block.Block;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Vec3d;
 import net.minecraft.util.math.Vec3i;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.Vec3;
@@ -82,8 +83,8 @@ public class BlockBounds {
 				side == EnumFacing.NORTH ? minZ : maxZ);
 	}
 
-	public Vec3i random(BlockPos pos, double centered) {
-		return new Vec3i(
+	public Vec3d random(BlockPos pos, double centered) {
+		return new Vec3d(
 				pos.getX() + RandUtils.getBetweenCentered(minX, maxX, centered),
 				pos.getY() + RandUtils.getBetweenCentered(minY, maxY, centered),
 				pos.getZ() + RandUtils.getBetweenCentered(minZ, maxZ, centered));
